Add an integration test for the crawler
It verifies that tweets are crawled and written the database. The
contents of the tweets must be manually verified for "meaness". It is a
small sample size of 100 tweets, so there is still a lot of noise, but
possible improvements might be using a sum of meaness (more mean words
= more mean) and checking for duplicate tweets (re-tweeting causes the
same tweet to have different IDs).